+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog].
 ### Removed
 
 ### Fixed
+- AutoMergeSkinnedMesh now chooses the generated renderer GameObject layer from the most common source renderer layer [`#1749`](https://github.com/anatawa12/AvatarOptimizer/pull/1749)
+  - This prevents layer-based cameras, lights, and shader gimmicks work correctly in Play mode after optimization.
 - Merge Bone may introduce z-fighting `#1748`
   - This fixes compatibility problem with see through hair